Start building default gameboard
The default gameboard will serve as an example of how to build an updating board using helpers located in core.js. core.js will be a javascript API to the server-side API. In the future I'd like to create another stock gameboard which will demonstrate how to leverage Handlebars block expressions to build a board.